+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Add new config-code component
 - Update config to new Cleep components
 - Improve monitoring (show/hide widgets, disable tasks)
+- Use TaskFactory
 
 ## [2.2.0] - 2023-03-09
 ### Fixed